refactor(refactor): merge utils into shared ns
